Some minor fixes, and added traits to Player for reentrant access.

+++ include/database/CharacterInfo.h
@@ -15,7 +15,7 @@ struct CharacterInfo
 	boost::uint32_t CharacterGrade;
     boost::uint16_t ClanPoints;
     boost::uint8_t CharacterMarker;
-    boost::uint16_t CharacterLevel;
+    boost::uint8_t CharacterLevel;
     boost::uint8_t CharacterSex;
     boost::uint8_t CharacterHair;
     boost::uint8_t CharacterFace;
